pyproject.toml
[tool.poetry]
name = "docker-fastapi-ml-inference"
version = "0.1.0"
description = "machine learning models and FastAPI endpoints"
authors = ["@BrianSigafoos"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.92.0"
httpx = "^0.23.3"
joblib = "^1.2.0"
pandas = "^1.5.3"
psycopg = "^3.1.8"
psycopg-pool = "^3.1.6"
python-dotenv = "^0.21.1"
scikit-learn = "^1.2.1"
sqlalchemy = "^2.0.4"
uvicorn = "^0.20.0"

[tool.poetry.group.dev.dependencies]
black = {extras = ["jupyter"], version = "^23.1.0"}
ipykernel = "^6.21.2"
isort = "^5.12.0"
pre-commit = "^3.0.4"
sshtunnel = "^0.4.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.2.1"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
